Updated look of homepage
Needs to be converted so the homepage pulls from the database and puts
actual stories where they belong... this should be a good jump-off point
to get the site barely functional. also, images will be needed to be
added at somepoint. I'm going to create a new branch for development as
well for best practice.