Category displays now work + abstracted some functionality + implemented widgets in the side bar
Abstracted some common tasks to functions in functions.php. Also
implemented category.php which displays a category that is requested by
the user (either in the navbar or on the main page).
Lots of room to improve upon the layout (perhaps add thumbnails?) the
sidebars are now dynamic as well, and can be configured in the wordpress
main panel and can have widgets added/removed.